search nearby for cities, restaurants, gas, etc
google maps api constraints on autocomplete are limited to cities, regions, and establishments, so the restaurants, gas, and establishments search all have the same autocomplete. We will just have to search quickly in demo to make it look good, and maybe compare to autocomplete in cities search which is just limited to cities.